Improve on_connection_lost hook
Previously, the connection lost hook wasn't listening for back-channel
events, which was a problem. If a client were not taking an action, and
only waiting for events to be delivered, they would never see a
ZK::Exceptions::ConnectionLoss error raised. This patch fixes that by
adding an event listener for ZOO_EXPIRED_SESSION_STATE, and triggering
the callback.